+++ pkg/scte35/audio_descriptor.go
@@ -46,28 +46,6 @@ func (sd *AudioDescriptor) Tag() uint32 {
 	return TimeDescriptorTag
 }
 
-// String returns the description of this SpliceDescriptor.
-func (sd *AudioDescriptor) String() string {
-	buf := &bytes.Buffer{}
-
-	buf.WriteString(fmt.Sprintf("audio_descriptor() {\n"))
-	buf.WriteString(fmt.Sprintf("    splice_descriptor_tag: %#02x\n", sd.Tag()))
-	buf.WriteString(fmt.Sprintf("    descriptor_length: %d bytes\n", sd.length()))
-	buf.WriteString(fmt.Sprintf("    identifier: %s\n", CUEIASCII))
-	buf.WriteString(fmt.Sprintf("    audio_count: %d\n", len(sd.AudioChannels)))
-	for i, ac := range sd.AudioChannels {
-		buf.WriteString(fmt.Sprintf("    audio_channel[%d] {", i))
-		buf.WriteString(fmt.Sprintf("        component_tag: %d\n", ac.ComponentTag))
-		buf.WriteString(fmt.Sprintf("        iso_code: %s\n", ac.ISOCode))
-		buf.WriteString(fmt.Sprintf("        bit_stream_mode: %d\n", ac.BitStreamMode))
-		buf.WriteString(fmt.Sprintf("        num_channels: %d)\n", ac.NumChannels))
-		buf.WriteString(fmt.Sprintf("        full_srvc_audio: %v\n", ac.FullSrvcAudio))
-		buf.WriteString(fmt.Sprintf("    }\n"))
-	}
-	buf.WriteString(fmt.Sprintf("}\n"))
-	return string(buf.Bytes())
-}
-
 // decode updates this SpliceDescriptor from binary.
 func (sd *AudioDescriptor) decode(b []byte) error {
 	r := iobit.NewReader(b)
@@ -124,6 +102,27 @@ func (sd *AudioDescriptor) length() int {
 		length += sd.AudioChannels[i].length() * 8
 	}
 	return length / 8
+}
+
+// table returns the tabular description of this SpliceDescriptor.
+func (sd *AudioDescriptor) table(prefix, indent string) string {
+	var buf bytes.Buffer
+	_, _ = fmt.Fprintf(&buf, prefix+"audio_descriptor() {\n")
+	_, _ = fmt.Fprintf(&buf, prefix+indent+"splice_descriptor_tag: %#02x\n", sd.Tag())
+	_, _ = fmt.Fprintf(&buf, prefix+indent+"descriptor_length: %d bytes\n", sd.length())
+	_, _ = fmt.Fprintf(&buf, prefix+indent+"identifier: %s\n", CUEIASCII)
+	_, _ = fmt.Fprintf(&buf, prefix+indent+"audio_count: %d\n", len(sd.AudioChannels))
+	for i, ac := range sd.AudioChannels {
+		_, _ = fmt.Fprintf(&buf, prefix+indent+"audio_channel[%d] {", i)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+indent+"component_tag: %d\n", ac.ComponentTag)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+indent+"iso_code: %s\n", ac.ISOCode)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+indent+"bit_stream_mode: %d\n", ac.BitStreamMode)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+indent+"num_channels: %d)\n", ac.NumChannels)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+indent+"full_srvc_audio: %v\n", ac.FullSrvcAudio)
+		_, _ = fmt.Fprintf(&buf, prefix+indent+"}\n")
+	}
+	_, _ = fmt.Fprintf(&buf, prefix+"}\n")
+	return string(buf.Bytes())
 }
 
 // AudioChannel collects the audio PID details.