+++ pkg/scte35/splice_insert.go
@@ -17,9 +17,9 @@
 package scte35
 
 import (
+	"bytes"
 	"encoding/xml"
 	"fmt"
-	"strings"
 
 	"github.com/bamiaux/iobit"
 )
@@ -52,49 +52,47 @@ func (cmd *SpliceInsert) Type() uint32 {
 	return SpliceInsertType
 }
 
-// String returns the table description of this splice_insert.
-func (cmd *SpliceInsert) String() string {
-	var buf strings.Builder
-
-	buf.WriteString(fmt.Sprintf("splice_insert() {\n"))
-	buf.WriteString(fmt.Sprintf("    splice_event_id: %d\n", cmd.SpliceEventID))
-	buf.WriteString(fmt.Sprintf("    splice_event_cancel_indicator: %v\n", cmd.SpliceEventCancelIndicator))
+// table returns the tabular description of this splice_insert.
+func (cmd *SpliceInsert) table(prefix, indent string) string {
+	var b bytes.Buffer
+	_, _ = fmt.Fprintf(&b, prefix+"splice_insert() {\n")
+	_, _ = fmt.Fprintf(&b, prefix+indent+"splice_event_id: %d\n", cmd.SpliceEventID)
+	_, _ = fmt.Fprintf(&b, prefix+indent+"splice_event_cancel_indicator: %v\n", cmd.SpliceEventCancelIndicator)
 	if !cmd.SpliceEventCancelIndicator {
-		buf.WriteString(fmt.Sprintf("    out_of_network_indicator: %v\n", cmd.OutOfNetworkIndicator))
-		buf.WriteString(fmt.Sprintf("    program_splice_flag: %v\n", cmd.programSpliceFlag()))
-		buf.WriteString(fmt.Sprintf("    duration_flag: %v\n", cmd.durationFlag()))
-		buf.WriteString(fmt.Sprintf("    splice_immediate_flag: %v\n", cmd.SpliceImmediateFlag))
+		_, _ = fmt.Fprintf(&b, prefix+indent+"out_of_network_indicator: %v\n", cmd.OutOfNetworkIndicator)
+		_, _ = fmt.Fprintf(&b, prefix+indent+"program_splice_flag: %v\n", cmd.programSpliceFlag())
+		_, _ = fmt.Fprintf(&b, prefix+indent+"duration_flag: %v\n", cmd.durationFlag())
+		_, _ = fmt.Fprintf(&b, prefix+indent+"splice_immediate_flag: %v\n", cmd.SpliceImmediateFlag)
 		if cmd.programSpliceFlag() && !cmd.SpliceImmediateFlag {
-			buf.WriteString(fmt.Sprintf("    time_specified_flag: %v\n", cmd.timeSpecifiedFlag()))
+			_, _ = fmt.Fprintf(&b, prefix+indent+"time_specified_flag: %v\n", cmd.timeSpecifiedFlag())
 			if cmd.timeSpecifiedFlag() {
-				buf.WriteString(fmt.Sprintf("    pts_time: %d ticks (%s)\n", *cmd.Program.SpliceTime.PTSTime, TicksToDuration(*cmd.Program.SpliceTime.PTSTime)))
+				_, _ = fmt.Fprintf(&b, prefix+indent+"pts_time: %d ticks (%s)\n", *cmd.Program.SpliceTime.PTSTime, TicksToDuration(*cmd.Program.SpliceTime.PTSTime))
 			}
 		}
 		if !cmd.programSpliceFlag() {
-			buf.WriteString(fmt.Sprintf("    component_count: %d\n", len(cmd.Components)))
-			for j, c := range cmd.Components {
-				buf.WriteString(fmt.Sprintf("    component[%d] {\n", j))
-				buf.WriteString(fmt.Sprintf("        component_tag: %d\n", c.Tag))
+			_, _ = fmt.Fprintf(&b, prefix+indent+"component_count: %d\n", len(cmd.Components))
+			for i, c := range cmd.Components {
+				_, _ = fmt.Fprintf(&b, prefix+indent+"component[%d] {\n", i)
+				_, _ = fmt.Fprintf(&b, prefix+indent+indent+"component_tag: %d\n", c.Tag)
 				if !cmd.SpliceImmediateFlag {
-					buf.WriteString(fmt.Sprintf("        time_specified_flag: %v\n", c.timeSpecifiedFlag()))
+					_, _ = fmt.Fprintf(&b, prefix+indent+indent+"time_specified_flag: %v\n", c.timeSpecifiedFlag())
 					if c.timeSpecifiedFlag() {
-						buf.WriteString(fmt.Sprintf("        pts_time: %d ticks (%s)\n", *c.SpliceTime.PTSTime, TicksToDuration(*c.SpliceTime.PTSTime)))
+						_, _ = fmt.Fprintf(&b, prefix+indent+indent+"pts_time: %d ticks (%s)\n", *c.SpliceTime.PTSTime, TicksToDuration(*c.SpliceTime.PTSTime))
 					}
 				}
-				buf.WriteString(fmt.Sprintf("    }\n"))
+				_, _ = fmt.Fprintf(&b, prefix+indent+"}\n")
 			}
 		}
 		if cmd.durationFlag() {
-			buf.WriteString(fmt.Sprintf("    auto_return: %v\n", cmd.BreakDuration.AutoReturn))
-			buf.WriteString(fmt.Sprintf("    duration: %d ticks (%s)\n", cmd.BreakDuration.Duration, TicksToDuration(cmd.BreakDuration.Duration)))
+			_, _ = fmt.Fprintf(&b, prefix+indent+"auto_return: %v\n", cmd.BreakDuration.AutoReturn)
+			_, _ = fmt.Fprintf(&b, prefix+indent+"duration: %d ticks (%s)\n", cmd.BreakDuration.Duration, TicksToDuration(cmd.BreakDuration.Duration))
 		}
-		buf.WriteString(fmt.Sprintf("    unique_program_id: %d\n", cmd.UniqueProgramID))
-		buf.WriteString(fmt.Sprintf("    avail_num: %d\n", cmd.AvailNum))
-		buf.WriteString(fmt.Sprintf("    avails_expected: %d\n", cmd.AvailsExpected))
-		buf.WriteString(fmt.Sprintf("}\n"))
+		_, _ = fmt.Fprintf(&b, prefix+indent+"unique_program_id: %d\n", cmd.UniqueProgramID)
+		_, _ = fmt.Fprintf(&b, prefix+indent+"avail_num: %d\n", cmd.AvailNum)
+		_, _ = fmt.Fprintf(&b, prefix+indent+"avails_expected: %d\n", cmd.AvailsExpected)
 	}
-
-	return buf.String()
+	_, _ = fmt.Fprintf(&b, prefix+"}\n")
+	return b.String()
 }
 
 // decode a binary splice_insert.