+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.0] - 2021-07-06
+
+### Added
+
+* `SpliceInfoSection.Table(prefix, indent)` added to allow formatting the 
+  tabular signal details.
+
+* Add missing spec and schema to `/docs` 
+
 ## [1.0.0] - 2021-07-01
 
 ### Notes